backend/requirements.txt
flask
accelerate
tokenizers
transformers
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

sqlalchemy
psycopg2-binary
flask-cors
python-dotenv
